README.md
This poject presents a bi-directional cognition interaction system with metaverse space via the wearable integrated interface introduced to this project. In general, our exploration achievements mainly involve in three aspects. Firstly, we created the metaverse space with Unity3D and HoudiniFX and connected them to hardware toolkits, thus actualizing the access to bi-directional cognition interaction. Secondly, we designed and built the shape-changing wearable integrated interface, realizing the function of multi-sensory perception. Last but not least, the latent living process embodied in human , such as brainwave, was introduced to the input sensing, enabling the wearer to communicate with the metaverse space through their daily behavioral habits in the physical world. In short, this interaction system provides the possibilities where people behave as if the metaverse space is the physical one, as well as future application for this technology. My role: Concept, Methodology; Technique solution structure and part of programming and fabrication.